refactor: move `Tile` class out of `level.py` to `tile.py`
